[+] tensor: the ops models kept re-implementing
Norms: layernorm_with/rms_norm_with builders with affine weight and bias,
normalizing in f32 and rounding once for low-precision inputs. Convs:
conv1d builder (stride, padding, dilation, groups) and batchnorm taking
running_var + eps so BN folding stays in the graph. Attention:
rope_table, split_heads/merge_heads, causal_mask, sequence_mask,
repeat_interleave, masked_mean, take_index, and key_padding_mask /
enable_gqa on the SDPA builder. Shape: narrow, unfold over pool,
pool_with defaults, upsample(_with), slice_with without the i64::MAX
sentinel, numel_sint, isize axes everywhere, symbolic non-reduced dims
through mean/var/std/argmax/cumsum/chunk/triu/unflatten. scatter_nd
takes a ScatterReduction enum; the bitwise_* aliases are gone.

+++ onnx/src/registry/transformer.rs
@@ -142,7 +142,7 @@ pub(crate) fn op_attention_onnx(inputs: &[Option<Tensor>], attrs: &mut Attrs) ->
     if is_causal {
         let past_seq_len = past_key.map(|pk| pk.dim_const(2)).transpose()?.unwrap_or(0);
         let q_len = q.dim_const(-2)?;
-        let causal = Tensor::full(&[q_len, full_k_len], true, DType::Bool).tril(past_seq_len as i64)?;
+        let causal = Tensor::full(&[q_len, full_k_len], true, DType::Bool).tril(past_seq_len as isize)?;
         let neg_inf = Tensor::const_(f64::NEG_INFINITY, q_dtype.clone());
         scores = scores.where_(&causal, &neg_inf)?;
     }
@@ -480,7 +480,7 @@ pub(crate) fn op_attention_contrib(inputs: &[Option<Tensor>], attrs: &mut Attrs)
     // Unidirectional causal mask
     if unidirectional {
         let causal =
-            Tensor::full(&[seq_len, total_seq], true, DType::Bool).tril((total_seq as i64) - (seq_len as i64))?;
+            Tensor::full(&[seq_len, total_seq], true, DType::Bool).tril(total_seq as isize - seq_len as isize)?;
         let filter = Tensor::const_(mask_filter_value, q_dtype.clone());
         let zero = Tensor::const_(0.0f64, q_dtype.clone());
         let causal_additive = zero.where_(&causal, &filter)?;
